requirements.txt
argparse
grpcio
grpcio-tools
protobuf
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3

